Fix layout for profile display, improve styling of rockets
